bug fixed

+++ src/main/java/com/usharesoft/basicmaven/App.java
@@ -14,7 +14,7 @@ public class App {
         int subtract = calculator.subtract(5, 10);
         System.out.println("5 - 10 = " + subtract);
 
-        int divide = calculator.divide(7, 2);
+        float divide = calculator.divide(7, 2);
         System.out.println("7 / 2 = " + divide);
 
         int multiply = calculator.multiply(5, 10);
